refactor(publisher,agent): extract shared keyed-lock + inline the serve-skip predicate (#1534 review r3)
Two maintainability follow-ups from the third review pass, both net-negative LOC.

1. Extract the gate-based multi-key lock. The agents/_meta retention module carried
   a private copy of the same gate-lock idiom used by dkg-publisher.ts /
   workspace-handler.ts. Lifted it into one shared helper `withKeyedLocks(lockMap,
   keys, fn)` (keyed-lock.ts); retention now calls it with its own key map, so the
   domain file expresses only insert/prune policy and there is a single lock
   implementation to audit. Per-individual-root serialization, disjoint-root
   concurrency, insert-first ordering, and swallowed post-insert prune failure are
   unchanged (agents-meta-bound A2 overlapping-root guard still green). The two
   PRE-EXISTING private copies (dkg-publisher/workspace-handler) sit on the
   concurrency-critical publish/share path; adopting the shared helper there is a
   clean mechanical follow-up kept OUT of this agents-meta PR to avoid widening its
   blast radius.

2. Inline the serve-skip predicate. Deleted the `createAgentsDurableMetaWithholdPredicate`
   currying factory; `start()` now calls the pure, unit-tested `shouldWithholdAgentsDurableMeta`
   resolver DIRECTLY from the injected predicate, with a fresh
   `process.env.DKG_SERVE_AGENTS_META` read at the lifecycle boundary. One concept
   instead of two; the policy module is back to two pure functions. Removed the
   now-redundant serve-wiring test (it only exercised the deleted factory) — the
   resolver's precedence is covered by agents-meta-policy.test.ts and the responder
   honoring the predicate by sync-responder-agents-meta-serve-skip.test.ts.

Tests: publisher full suite 1449 pass / 6 skip (agents-meta-bound 16/16 incl. A2);
agent agents-meta-policy 11/11 + sync-responder-agents-meta-serve-skip 4/4. Full
core→agent→cli typecheck clean.
